requirements.txt
streamlit
torch
demucs
openai-whisper
pydub
